+++ HISTORY.md
@@ -1,4 +1,7 @@
 * 2021-03-17
 	* First release.
 
+* 2021-04-14
+	* Dependeny to pypine removed
+	* README.md updated